Added piece-square tables

+++ src/edu/kingsu/SoftwareEngineering/Chess/Players/Move.java
@@ -7,10 +7,10 @@ import edu.kingsu.SoftwareEngineering.Chess.Board.Pieces.*;
  * A move class to hold start and end locations as well as the piece to move
  */
 public class Move {
-    public final int score;
+    public final double score;
     public final Piece piece;
     public final BoardLocation start, end;
-    public Move(Piece piece, BoardLocation start, BoardLocation end, int score) {
+    public Move(Piece piece, BoardLocation start, BoardLocation end, double score) {
         this.piece = piece;
         this.start = start;
         this.end = end;
